Create an object that handles both SID mapping and pseudonymization
