+++ src/main/java/no/ssb/dlp/pseudo/service/pseudo/PseudoField.java
@@ -2,29 +2,43 @@ package no.ssb.dlp.pseudo.service.pseudo;
 
 import lombok.*;
 import lombok.experimental.SuperBuilder;
+import lombok.extern.log4j.Log4j2;
+import no.ssb.dlp.pseudo.core.field.FieldDescriptor;
+import no.ssb.dlp.pseudo.core.field.FieldPseudonymizer;
 import no.ssb.dlp.pseudo.core.func.PseudoFuncRule;
 import no.ssb.dlp.pseudo.core.tink.model.EncryptedKeysetWrapper;
 
+import java.time.Duration;
+import java.time.Instant;
+import java.util.ArrayList;
 import java.util.List;
 
 /**
  * Represents a field to be pseudonymized.
  */
-public class PseudoField extends AbstractPseudoField<ResponsePseudoField> {
+@Log4j2
+@Data
+public class PseudoField {
     @Getter(AccessLevel.PROTECTED)
     private static final String DEFAULT_PSEUDO_FUNC = "daead(keyId=ssb-common-key-1)";
+
+    protected String name;
+    protected List<String> values;
+    protected PseudoConfig pseudoConfig;
+    private List<String> sidValues;
 
     /**
      * Constructs a {@code PseudoField} object with the specified name, values, keyset, pseudoConfig. If no keyset is supplied
      * a default pseudo configuration is used.
      *
-     * @param name   The name of the field.
-     * @param values The values of the field.
+     * @param name       The name of the field.
+     * @param values     The values of the field.
      * @param pseudoFunc The pseudo function definition.
-     * @param keyset The encrypted keyset to be used for pseudonymization.
+     * @param keyset     The encrypted keyset to be used for pseudonymization.
      */
     public PseudoField(String name, List<String> values, String pseudoFunc, EncryptedKeysetWrapper keyset) {
-        super(name, values);
+        this.name = name;
+        this.values = values;
 
         pseudoConfig = new PseudoConfig();
 
@@ -47,6 +61,23 @@ public class PseudoField extends AbstractPseudoField<ResponsePseudoField> {
     public ResponsePseudoField pseudonymizeThenGetResponseField(RecordMapProcessorFactory recordProcessorFactory) {
         List<String> encryptedValues = pseudonymize(values, recordProcessorFactory);
         return this.getResponseField(encryptedValues);
+    }
+
+
+    protected List<String> pseudonymize(List<String> values, RecordMapProcessorFactory recordProcessorFactory) {
+        Instant startTime = Instant.now();
+
+        FieldPseudonymizer fieldPseudonymizer = recordProcessorFactory.newFieldPseudonymizer(this.getPseudoConfig().getRules(), RecordMapProcessorFactory.pseudoKeysetsOf(this.getPseudoConfig().getKeysets()));
+
+        ArrayList<String> encryptedValues = new ArrayList<>();
+
+        values.stream().map(value -> fieldPseudonymizer.pseudonymize(new FieldDescriptor(this.getName()), value)).forEach(result -> encryptedValues.add(result));
+
+        Instant endTime = Instant.now();
+        Duration duration = Duration.between(startTime, endTime);
+        log.info("Pseudonymizing field '{}' took {} milliseconds.", this.getName(), duration.toMillis());
+
+        return encryptedValues;
     }
